feat: design and implement HeaderSection component
