+++ src/app/globals.css
@@ -3,8 +3,11 @@
 @tailwind utilities;
 
 :root {
+  font-size: 16px;
   --background: #ffffff;
   --foreground: #171717;
+  --color-text: #f5f3ee;
+  --color-bg: #0e0d0e;
 }
 
 @media (prefers-color-scheme: dark) {
@@ -15,7 +18,54 @@
 }
 
 body {
-  color: var(--foreground);
-  background: var(--background);
-  font-family: Arial, Helvetica, sans-serif;
+  width: 100vw;
+  overflow-x: hidden;
+}
+
+body {
+  margin: 0;
+  color: var(--color-text);
+  background-color: var(--color-bg);
+  font-variation-settings:
+    'slnt' 0,
+    'wdth' 100,
+    'wght' 450,
+    'INKT' 0;
+  font-weight: 450;
+  line-height: 1.2;
+}
+
+/* Page Loader */
+.js .loading::before,
+.js .loading::after {
+  content: '';
+  position: fixed;
+  z-index: 1000;
+}
+
+.js .loading::before {
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: var(--color-bg);
+}
+
+.js .loading::after {
+  top: 50%;
+  left: 50%;
+  width: 60px;
+  height: 60px;
+  margin: -30px 0 0 -30px;
+  border-radius: 50%;
+  opacity: 0.4;
+  background: var(--color-link);
+  animation: loaderAnim 0.7s linear infinite alternate forwards;
+}
+
+@keyframes loaderAnim {
+  to {
+    opacity: 1;
+    transform: scale3d(0.5, 0.5, 1);
+  }
 }
