+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './core'
-export * from './layout'
+export * from './layouts'
+export * from './sections'
