--- src/components/layout/Header.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export const Header = () => {
-  return <div>Hello World</div>
-}
